+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.1.1
+
+* Major API re-design; no backwards compatibility.
+* `CurrencyParser::Parser` is now a `Class`.
+
 # v0.1.0
 
 * Initial version.